+++ ackermann_steering_controller/src/ackermann_steering_controller.cpp
@@ -12,6 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <limits>
+
 #include "ackermann_steering_controller/ackermann_steering_controller.hpp"
 
 namespace ackermann_steering_controller
@@ -31,6 +33,8 @@ controller_interface::CallbackReturn AckermannSteeringController::configure_odom
 {
   ackermann_params_ = ackermann_param_listener_->get_params();
 
+  // TODO(anyone): Remove deprecated parameters
+  // START OF DEPRECATED
   if (ackermann_params_.front_wheels_radius > 0.0)
   {
     RCLCPP_WARN(
@@ -51,23 +55,47 @@ controller_interface::CallbackReturn AckermannSteeringController::configure_odom
   {
     RCLCPP_WARN(
       get_node()->get_logger(),
-      "DEPRECATED parameter 'front_wheel_track', set 'traction_wheel_track' instead");
-    ackermann_params_.traction_wheel_track = ackermann_params_.front_wheel_track;
+      "DEPRECATED parameter 'front_wheel_track', set 'traction_wheel_track' or "
+      "'steering_wheel_track' instead");
+    if (params_.front_steering)
+    {
+      ackermann_params_.steering_wheel_track = ackermann_params_.front_wheel_track;
+    }
+    else
+    {
+      ackermann_params_.traction_wheel_track = ackermann_params_.front_wheel_track;
+    }
   }
 
   if (ackermann_params_.rear_wheel_track > 0.0)
   {
     RCLCPP_WARN(
       get_node()->get_logger(),
-      "DEPRECATED parameter 'rear_wheel_track', set 'traction_wheel_track' instead");
-    ackermann_params_.traction_wheel_track = ackermann_params_.rear_wheel_track;
+      "DEPRECATED parameter 'rear_wheel_track', set 'traction_wheel_track' or "
+      "'steering_wheel_track' instead");
+    if (params_.front_steering)
+    {
+      ackermann_params_.traction_wheel_track = ackermann_params_.rear_wheel_track;
+    }
+    else
+    {
+      ackermann_params_.steering_wheel_track = ackermann_params_.rear_wheel_track;
+    }
+  }
+  // END OF DEPRECATED
+
+  if (ackermann_params_.steering_wheel_track <= std::numeric_limits<double>::epsilon())
+  {
+    ackermann_params_.steering_wheel_track = ackermann_params_.traction_wheel_track;
   }
 
   const double traction_wheels_radius = ackermann_params_.traction_wheels_radius;
   const double traction_wheel_track = ackermann_params_.traction_wheel_track;
+  const double steering_wheel_track = ackermann_params_.steering_wheel_track;
   const double wheelbase = ackermann_params_.wheelbase;
 
-  odometry_.set_wheel_params(traction_wheels_radius, wheelbase, traction_wheel_track);
+  odometry_.set_wheel_params(
+    traction_wheels_radius, wheelbase, steering_wheel_track, traction_wheel_track);
   odometry_.set_odometry_type(steering_odometry::ACKERMANN_CONFIG);
 
   set_interface_numbers(NR_STATE_ITFS, NR_CMD_ITFS, NR_REF_ITFS);